requirement.txt
tk
matplotlib
pandas
sklearn
tensorflow
opencv-python
Pillow
numpy
pyautogui